WIP bugfixes
