+++ modules/app_store/app_store/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "app_store"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
